refactor: Browser index cleanup
